Revert "씬 분리"
This reverts commit 4dbfd671921eda94bcf78e783769b95c89246b20.

+++ CameraManager.cpp
@@ -9,7 +9,7 @@ CameraManager::~CameraManager()
 {
 }
 
-HRESULT CameraManager::init(int mapSizeX, float mapSizeY)
+HRESULT CameraManager::init(int mapSizeX, int mapSizeY)
 {
     _posX = WINSIZEX / 2;
     _posY = WINSIZEY / 2;
